requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf

#embeddings
sentence-transformers

#vector database
faiss-cpu
chromadb

#llm
langchain-groq
python-dotenv
streamlit